+++ packages/qwik/src/validation/maxFiles.ts
@@ -1,9 +1,12 @@
 import { $, type NoSerialize, type QRL } from '@builder.io/qwik';
+import type { MaybeValue } from '../types';
 
-type Value = NoSerialize<Blob[]> | NoSerialize<File[]> | null | undefined;
+type Value = MaybeValue<NoSerialize<Blob[]> | NoSerialize<File[]>>;
 
 /**
  * Creates a validation functions that validates the size of a file list.
+ *
+ * @deprecated Please use `maxLength` instead.
  *
  * @param requirement The maximum size.
  * @param error The error message.