+++ packages/qwik/src/validation/maxTotalSize.ts
@@ -1,6 +1,7 @@
 import { $, type NoSerialize, type QRL } from '@builder.io/qwik';
+import type { MaybeValue } from '../types';
 
-type Value = NoSerialize<Blob[]> | NoSerialize<File[]> | null | undefined;
+type Value = MaybeValue<NoSerialize<Blob[]> | NoSerialize<File[]>>;
 
 /**
  * Creates a validation functions that validates total file size of a file list.