+++ packages/qwik/src/validation/minRange.ts
@@ -1,9 +1,10 @@
 import { $, type QRL } from '@builder.io/qwik';
+import type { MaybeValue } from '../types';
 
-type Value = string | Date | null | undefined;
+type Value = MaybeValue<string | number | Date>;
 
 /**
- * Creates a validation functions that validates the range of a string or date.
+ * Creates a validation functions that validates the range of a string, number or date.
  *
  * @param requirement The minimum range.
  * @param error The error message.
@@ -11,8 +12,10 @@ type Value = string | Date | null | undefined;
  * @returns A validation function.
  */
 export function minRange(
-  requirement: string | Date,
+  requirement: string | number | Date,
   error: string
 ): QRL<(value: Value) => string> {
-  return $((value: Value) => (value && value < requirement ? error : ''));
+  return $((value: Value) =>
+    (value || value === 0) && value < requirement ? error : ''
+  );
 }